circular queue assert fix from LSQ unit
Adding the fix for assert from LSQ unit. This fix is already checked-in  into the public/gem5 repo.
'cpu-o3: Resolve circular buffer issue for LSQ'
Link: https://gem5-review.googlesource.com/c/public/gem5/+/58649

Reviewed-on: https://soc-sim-external-review.googlesource.com/c/gem5/gem5/+/8192
Tested-by: kokoro <noreply+kokoro@google.com>
Presubmit-Ready: Radhika Ranwadkar <rranwadkar@google.com>

+++ src/cpu/o3/lsq.cc
@@ -1411,7 +1411,7 @@ LSQ::HtmCmdRequest::finish(const Fault &fault, const RequestPtr &req,
 }
 
 Fault
-LSQ::read(LSQRequest* request, int load_idx)
+LSQ::read(LSQRequest* request, ssize_t load_idx)
 {
     assert(request->req()->contextId() == request->contextId());
     ThreadID tid = cpu->contextToThread(request->req()->contextId());
@@ -1420,7 +1420,7 @@ LSQ::read(LSQRequest* request, int load_idx)
 }
 
 Fault
-LSQ::write(LSQRequest* request, uint8_t *data, int store_idx)
+LSQ::write(LSQRequest* request, uint8_t *data, ssize_t store_idx)
 {
     ThreadID tid = cpu->contextToThread(request->req()->contextId());
